test(anyharness): split file authority coverage
