+++ anyharness/crates/anyharness-lib/src/adapters/files/mod.rs
@@ -4,4 +4,8 @@ pub mod service;
 pub mod types;
 
 #[cfg(test)]
+mod service_authority_tests;
+#[cfg(test)]
+mod service_test_support;
+#[cfg(test)]
 mod service_tests;
